消费记录的分页展示

+++ service/src/main/java/com/aalife/service/impl/CostGroupServiceImpl.java
@@ -8,6 +8,7 @@ import com.aalife.dao.entity.CostGroupUser;
 import com.aalife.dao.entity.CostUserRemark;
 import com.aalife.dao.entity.User;
 import com.aalife.dao.repository.CostDetailRepository;
+import com.aalife.dao.repository.CostGroupApprovalRepository;
 import com.aalife.dao.repository.CostGroupRepository;
 import com.aalife.dao.repository.CostGroupUserRepository;
 import com.aalife.dao.repository.CostUserRemarkRepository;
@@ -45,6 +46,8 @@ public class CostGroupServiceImpl implements CostGroupService {
     @Autowired
     private CostDetailRepository costDetailRepository;
     @Autowired
+    private CostGroupApprovalRepository costGroupApprovalRepository;
+    @Autowired
     private WebContext webContext;
 
     @Override
@@ -76,10 +79,16 @@ public class CostGroupServiceImpl implements CostGroupService {
     @Override
     public void updateCostGroup(CostGroupBo costGroupBo) {
         String groupName = costGroupBo.getGroupName();
+        Integer groupId = costGroupBo.getGroupNo();
         if (StringUtils.isEmpty(groupName)){
             throw new BizException("账单名不能为空");
         }
-        costGroupRepository.updateCostGroup(costGroupBo.getGroupNo(), groupName);
+        CostGroup costGroup = costGroupRepository.findGroupById(groupId);
+        if (costGroup == null){
+            throw new BizException("未查询到账单");
+        }
+        costGroup.setGroupName(groupName);
+        costGroupRepository.save(costGroup);
     }
 
     @Override
@@ -115,6 +124,23 @@ public class CostGroupServiceImpl implements CostGroupService {
         User user = webContext.getCurrentUser();
         Integer userId = user.getUserId();
         // todo 推出前需要判断，所有消费记录必须结算，必须要有管理员
+        // 校验是否有管理员，除了当前用户
+        List<CostGroupUser> costGroupUsers = costGroupUserRepository.findCostGroupByGroup(groupId);
+        boolean hasAnotherAdmin = false;
+        for (CostGroupUser costGroupUser : costGroupUsers){
+            // 存在除去当前用户的管理员
+            if (costGroupUser.getAdmin().equals('Y') && !costGroupUser.getUser().getUserId().equals(userId)){
+                hasAnotherAdmin = true;
+            }
+        }
+        if (!hasAnotherAdmin){
+            throw new BizException("退出该账单前需指定新的管理员");
+        }
+        // 校验该用户的账单是否已经结算
+        BigDecimal groupCost = costDetailRepository.findUnCleanTotalCostByGroup(groupId);
+        if (groupCost != null || groupCost.intValue() != 0){
+            throw new BizException("所在分组未结算，请先结算后再退出或联系账单管理员删除");
+        }
         costGroupUserRepository.deleteCostGroupUser(userId, groupId, userId);
     }
 
@@ -127,7 +153,14 @@ public class CostGroupServiceImpl implements CostGroupService {
     public void deleteCostGroupUser(Integer groupId, Integer userId) {
         User user = webContext.getCurrentUser();
         Integer deleteId = user.getUserId();
+        if (userId.equals(deleteId)){
+            throw new BizException("不能删除自己，可以选择退出操作或联系另一管理员");
+        }
         // todo 推出前需要判断，所有消费记录必须结算，必须要有管理员
+        BigDecimal userCost = costDetailRepository.findUnCleanTotalCostByUserAndGroup(groupId, userId);
+        if (userCost != null && userCost.intValue() != 0){
+            throw new BizException("该用户有未结算的记录");
+        }
         costGroupUserRepository.deleteCostGroupUser(userId, groupId, deleteId);
     }
 
@@ -136,6 +169,9 @@ public class CostGroupServiceImpl implements CostGroupService {
         CostGroupOverviewBo costGroupOverviewBo = new CostGroupOverviewBo();
         // 设置组信息
         CostGroup costGroup = costGroupRepository.findGroupById(groupId);
+        if (costGroup == null){
+            throw new BizException("未查询到账单");
+        }
         CostGroupBo costGroupBo = new CostGroupBo();
         costGroupBo.setGroupNo(costGroup.getGroupId());
         costGroupBo.setGroupName(costGroup.getGroupName());
@@ -177,13 +213,17 @@ public class CostGroupServiceImpl implements CostGroupService {
         costGroupOverviewBo.setGroupTotalCost(groupTotalCost);
         // 设置消费平均消费和差价
         BigDecimal userCount = new BigDecimal(costGroupUserBos.size());
-        BigDecimal averageCost = groupTotalCost.divide(userCount);
+        // 指定精确的位数，否者报错
+        BigDecimal averageCost = groupTotalCost.divide(userCount, 2);
         for (CostGroupUserBo costGroupBoTemp : costGroupUserBos){
             BigDecimal costMoney = costGroupBoTemp.getTotalCost();
             costGroupBoTemp.setAverageCost(averageCost);
             costGroupBoTemp.setLeftCost(averageCost.subtract(costMoney));
         }
         costGroupOverviewBo.setCostUsers(costGroupUserBos);
+        // 设置待接受的数量
+        Integer count = costGroupApprovalRepository.getNotApproveUserCount(groupId);
+        costGroupOverviewBo.setNotApproveCount(count == null ? 0 : count);
         return costGroupOverviewBo;
     }
